+++ actionmailbox/app/controllers/rails/conductor/action_mailbox/inbound_emails_controller.rb
@@ -22,7 +22,7 @@ module Rails
       def new_mail
         Mail.new(params.require(:mail).permit(:from, :to, :cc, :bcc, :in_reply_to, :subject, :body).to_h).tap do |mail|
           params[:mail][:attachments].to_a.each do |attachment|
-            mail.attachments[attachment.original_filename] = { filename: attachment.path, content_type: attachment.content_type }
+            mail.add_file(filename: attachment.path, content: attachment.read)
           end
         end
       end