+++ activerecord/lib/active_record/migration/compatibility.rb
@@ -36,9 +36,7 @@ module ActiveRecord
       class V5_1 < V5_2
         def change_column(table_name, column_name, type, options = {})
           if adapter_name == "PostgreSQL"
-            clear_cache!
-            sql = connection.send(:change_column_sql, table_name, column_name, type, options)
-            execute "ALTER TABLE #{quote_table_name(table_name)} #{sql}"
+            super(table_name, column_name, type, options.except(:default, :null, :comment))
             change_column_default(table_name, column_name, options[:default]) if options.key?(:default)
             change_column_null(table_name, column_name, options[:null], options[:default]) if options.key?(:null)
             change_column_comment(table_name, column_name, options[:comment]) if options.key?(:comment)