+++ actionmailbox/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 6.0.0.beta1 (January 18, 2019) ##
+
 *   Added to Rails.
 
     *DHH*