+++ actionview/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 6.0.0.beta1 (January 18, 2019) ##
+
 *   Remove deprecated `image_alt` helper.
 
     *Rafael Mendonça França*
@@ -8,7 +10,8 @@
     *Genadi Samokovarov*
 
 *   Fix UJS permanently showing disabled text in a[data-remote][data-disable-with] elements within forms.
-    Fixes #33889
+
+    Fixes #33889.
 
     *Wolfgang Hobmaier*
 
@@ -20,7 +23,7 @@
     <%= link_to 'Remote', remote_path, class: 'remote', remote: true, data: { type: :json } %>
     ```
 
-    Fixes #34541
+    Fixes #34541.
 
     *Wolfgang Hobmaier*
 
@@ -37,7 +40,7 @@
 
     Calling `word_wrap` should not trim the indents on the first and last lines.
 
-    Fixes #34487
+    Fixes #34487.
 
     *Lyle Mullican*