+++ activesupport/CHANGELOG.md
@@ -24,6 +24,8 @@
     
     *Abhishek Sarkar*
     
+## Rails 6.0.0.beta1 (January 18, 2019) ##
+
 *   Remove deprecated `Module#reachable?` method.
 
     *Rafael Mendonça França*